+++ inertia/components/features/profile/UploadAvatarModal.tsx
@@ -1,4 +1,5 @@
 import { ImageUp } from "lucide-react";
+import { useEffect, useRef } from "react";
 import Modal, { type BaseModalProps } from "~/components/ui/modals/Modal";
 import UploadAvatarForm from "./UploadAvatarForm";
 
@@ -8,6 +9,14 @@ export default function UploadAvatarModal({
 	open,
 	onClose,
 }: UploadAvatarModalProps) {
+	const formRef = useRef<{ resetForm: () => void }>(null);
+
+	useEffect(() => {
+		if (!open) {
+			formRef.current?.resetForm();
+		}
+	}, [open]);
+
 	return (
 		<Modal
 			title="Avatar"
@@ -16,7 +25,7 @@ export default function UploadAvatarModal({
 			onClose={onClose}
 			icon={<ImageUp size="30" />}
 		>
-			<UploadAvatarForm onClose={onClose} />
+			<UploadAvatarForm ref={formRef} onClose={onClose} />
 		</Modal>
 	);
 }